src/master_station/requirements.txt
pyyaml
python-can
canstruct
urwid
